requirements.txt
# Core Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# OpenAI Integration
openai>=1.0.0

# Vector Database
chromadb>=0.4.0

# HTTP Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# CLI Enhancements
rich>=13.0.0

# Web Scraping & Content Extraction
trafilatura>=1.6.0
beautifulsoup4>=4.12.0
youtube-transcript-api>=0.6.0

# Retry Logic
tenacity>=8.2.0

# Configuration Files
pyyaml>=6.0

# System Utilities
psutil>=5.9.0

# Text Processing
langchain-text-splitters>=0.0.1
